Trying to get bottom sheet working
