+++ app/src/main/java/com/patrickchow/financeforsuccess/dataclass/ModalSheetInformation.kt
@@ -1,0 +1,6 @@
+package com.patrickchow.financeforsuccess.dataclass
+
+data class ModalSheetInformation(
+    val title: String,
+    val description: String,
+)
